fix(ci): remove obsolete remote execution gate tests
