+++ tests/test_remote_access_vibe_cloud.py
@@ -16,7 +16,7 @@ from config.v2_config import AgentsConfig, PlatformsConfig, RemoteAccessConfig, 
 from storage.db import create_sqlite_engine
 from storage.models import remote_access_authorizations
 from tests.ui_server_test_helpers import remote_session_cookie
-from vibe import remote_access
+from vibe import remote_access, ui_server
 from vibe import runtime
 
 
@@ -92,6 +92,64 @@ def _session_cookie(
         subject,
         session_claims=_session_claims(config, role=role),
     )
+
+
+def test_ra_tq_026_remote_status_uses_cf_ray_on_paired_public_host(monkeypatch, tmp_path):
+    monkeypatch.setenv("AVIBE_HOME", str(tmp_path))
+    config = _config()
+    config.save()
+    observed = []
+
+    def status(
+        loaded_config=None,
+        *,
+        client_colo=None,
+        client_access="local",
+        include_network_path=False,
+    ):
+        observed.append((loaded_config, client_colo, client_access, include_network_path))
+        return {"ok": True, "client_colo": client_colo}
+
+    monkeypatch.setattr(remote_access, "status", status)
+    with ui_server.app.test_request_context(
+        "/api/remote-access/status",
+        base_url="https://alex.avibe.bot",
+        headers={"CF-Ray": "9f1234567890abcd-SIN"},
+    ):
+        response = ui_server.remote_access_status()
+
+    assert response.status_code == 200
+    assert observed[0][0].remote_access.vibe_cloud.public_url == config.remote_access.vibe_cloud.public_url
+    assert observed[0][1:] == ("SIN", "remote", True)
+
+
+def test_ra_tq_026_remote_status_ignores_spoofed_cf_ray(monkeypatch, tmp_path):
+    monkeypatch.setenv("AVIBE_HOME", str(tmp_path))
+    config = _config()
+    config.save()
+    observed = []
+
+    def status(
+        loaded_config=None,
+        *,
+        client_colo=None,
+        client_access="local",
+        include_network_path=False,
+    ):
+        observed.append((loaded_config, client_colo, client_access, include_network_path))
+        return {"ok": True, "client_colo": client_colo}
+
+    monkeypatch.setattr(remote_access, "status", status)
+    with ui_server.app.test_request_context(
+        "/api/remote-access/status",
+        base_url="http://127.0.0.1:5123",
+        headers={"CF-Ray": "9f1234567890abcd-NRT"},
+    ):
+        response = ui_server.remote_access_status()
+
+    assert response.status_code == 200
+    assert observed[0][1:] == (None, "local", True)
+    assert config.remote_access.vibe_cloud.public_url == "https://alex.avibe.bot"
 
 
 def test_session_cookie_roundtrip() -> None:
